requirements.txt
amqp==2.5.1
astroid==2.2.5
billiard==3.6.1.0
celery==4.3.0
certifi==2019.9.11
chardet==3.0.4
Django==2.2.6
django-celery-beat==1.5.0
django-celery-results==1.1.2
django-timezone-field==3.1
djangorestframework==3.10.3
feedparser==5.2.1
idna==2.8
importlib-metadata==0.23
isort==4.3.21
kombu==4.6.5
lazy-object-proxy==1.4.1
mccabe==0.6.1
more-itertools==7.2.0
mysqlclient==1.4.4
pylint==2.3.1
python-crontab==2.3.9
python-dateutil==2.8.0
pytz==2019.3
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.6
vine==1.3.0
wrapt==1.11.2
zipp==0.6.0
